requirements.txt
psutil
py-cpuinfo
pynvml
PySide6
WMI
